Change size_to_validate from usize to u64 (#814)
Enforce that size_to_validate should be 64 bits wide, to allow
validating disks that are larger than 2**32 bytes.

+++ integration_tests/src/lib.rs
@@ -4080,7 +4080,7 @@ mod test {
                         ),
 
                     // validate half the volume
-                    size_to_validate: Some(total_size as u32 / 2),
+                    size_to_validate: Some(total_size as u64 / 2),
                 },
             )
             .await
